refactor: create BackupUtils and replaces use quant embedding store files during backup
